% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3vcov.R
\name{deviance}
\alias{deviance}
\alias{deviance.MIRT}
\alias{deviance.MGPCM}
\alias{deviance.MGGUM}
\alias{deviance.FCMIRT}
\alias{deviance.FCDCM}
\alias{deviance.FCGDINA}
\alias{deviance.FCGGUM}
\alias{deviance.TIRT}
\title{S3 Methods: deviance}
\usage{
\method{deviance}{MIRT}(object, ...)

\method{deviance}{MGPCM}(object, ...)

\method{deviance}{MGGUM}(object, ...)

\method{deviance}{FCMIRT}(object, ...)

\method{deviance}{FCDCM}(object, ...)

\method{deviance}{FCGDINA}(object, ...)

\method{deviance}{FCGGUM}(object, ...)

\method{deviance}{TIRT}(object, ...)
}
\arguments{
\item{object}{A fitted model object.}

\item{...}{Additional arguments (currently ignored).}
}
\value{
A numeric scalar: \eqn{-2 \log L}.
}
\description{
Returns the model deviance, defined as \eqn{-2 \ell} where \eqn{\ell}
is the marginal log-likelihood evaluated at the parameter estimates.
}
\section{Methods (by class)}{
\itemize{
\item \code{deviance(MIRT)}: Model deviance (\eqn{-2 \ell}) for \code{MIRT} objects.

\item \code{deviance(MGPCM)}: Model deviance for \code{MGPCM} objects.

\item \code{deviance(MGGUM)}: Model deviance for \code{MGGUM} objects.

\item \code{deviance(FCMIRT)}: Model deviance for \code{FCMIRT} objects.

\item \code{deviance(FCDCM)}: Model deviance for \code{FCDCM} objects.

\item \code{deviance(FCGDINA)}: Model deviance for \code{FCGDINA} objects.

\item \code{deviance(FCGGUM)}: Model deviance for \code{FCGGUM} objects.

\item \code{deviance(TIRT)}: Model deviance for \code{TIRT} objects.

}}
