% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools-mcmc.R
\name{extract_corr_stan}
\alias{extract_corr_stan}
\title{Extract correlation posterior summary from Stan MCMC output}
\usage{
extract_corr_stan(MCMC.obj, stan.sum, D)
}
\arguments{
\item{MCMC.obj}{Output of \code{rstan::extract()}.}

\item{stan.sum}{Output of \code{rstan::summary()$summary}.}

\item{D}{Number of dimensions.}
}
\value{
A list with components \code{est}, \code{se}, and \code{Rhat}.
}
\description{
Extract correlation posterior summary from Stan MCMC output
}
\keyword{internal}
