% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{free_corr_npar}
\alias{free_corr_npar}
\title{Degrees of freedom for a correlation matrix}
\usage{
free_corr_npar(D, include_corr = TRUE)
}
\arguments{
\item{D}{Number of dimensions.}

\item{include_corr}{Logical; if \code{FALSE} returns 0.}
}
\value{
Integer, the number of free parameters in a D x D correlation matrix.
}
\description{
Degrees of freedom for a correlation matrix
}
\keyword{internal}
