% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-utils.R
\name{generate_theta_mvn}
\alias{generate_theta_mvn}
\title{Generate latent trait vectors from a multivariate normal distribution}
\usage{
generate_theta_mvn(N, D, Corr)
}
\arguments{
\item{N}{Number of persons.}

\item{D}{Number of latent dimensions.}

\item{Corr}{A \eqn{D \times D} correlation matrix.}
}
\value{
An \eqn{N \times D} numeric matrix of latent trait values.
}
\description{
When \code{D > 1}, uses \code{MASS::mvrnorm} with the supplied correlation
matrix.  When \code{D == 1}, uses standard normal draws.
}
\keyword{internal}
