% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-utils.R
\name{get_precedent_patterns}
\alias{get_precedent_patterns}
\title{Find precedent (subset) patterns for monotonicity enforcement}
\usage{
get_precedent_patterns(patterns)
}
\arguments{
\item{patterns}{A binary matrix of attribute mastery patterns
(\eqn{L \times K}).}
}
\value{
A list of length \eqn{L}; the \eqn{l}-th element is an integer
vector of row indices corresponding to patterns that are proper
subsets of pattern \eqn{l}.
}
\description{
For each attribute mastery pattern, identifies which other patterns are
proper subsets (i.e., mastering a subset of the attributes).  Used to
enforce the GDINA monotonicity constraint: mastering more attributes
should never decrease the success probability.
}
\details{
Equivalent to \code{preloclist} in the GDINA R package.
}
\keyword{internal}
