% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{mggum_par_free_mask}
\alias{mggum_par_free_mask}
\title{Free-parameter mask for MGGUM models}
\usage{
mggum_par_free_mask(Q.matrix, length.poly, max_poly = max(length.poly))
}
\arguments{
\item{Q.matrix}{I x D Q-matrix.}

\item{length.poly}{Integer vector of category counts per item.}

\item{max_poly}{Maximum number of categories.}
}
\value{
An I x (2D + max_poly) logical matrix.
}
\description{
Free-parameter mask for MGGUM models
}
\keyword{internal}
