% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{mgpcm_par_free_mask}
\alias{mgpcm_par_free_mask}
\title{Free-parameter mask for MGPCM models}
\usage{
mgpcm_par_free_mask(Q.matrix, length.poly, max_poly = max(length.poly))
}
\arguments{
\item{Q.matrix}{I x D Q-matrix.}

\item{length.poly}{Integer vector of category counts per item.}

\item{max_poly}{Maximum number of categories (default \code{max(length.poly)}).}
}
\value{
An I x (D + max_poly) logical matrix.
}
\description{
Free-parameter mask for MGPCM models
}
\keyword{internal}
