% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{mirt_par_free_mask}
\alias{mirt_par_free_mask}
\title{Free-parameter mask for MIRT models}
\usage{
mirt_par_free_mask(model_type, Q.matrix)
}
\arguments{
\item{model_type}{Integer 2, 3, or 4.}

\item{Q.matrix}{I x D Q-matrix.}
}
\value{
An I x (D+3) logical matrix.
}
\description{
Free-parameter mask for MIRT models
}
\keyword{internal}
