% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.data.MIRT.R
\name{model.MIRT}
\alias{model.MIRT}
\title{Compute MIRT Item Response Probabilities}
\usage{
model.MIRT(theta, par)
}
\arguments{
\item{theta}{An \eqn{N \times D} matrix of latent trait values.}

\item{par}{An \eqn{I \times (D+3)} parameter matrix with columns
\code{a1..aD, b, c, d}, where \code{b} is the item difficulty.
The linear predictor is \eqn{\mathbf{a}_i'\boldsymbol{\theta}_j - b_i}.}
}
\value{
An \eqn{N \times I} matrix of response probabilities
\eqn{P(Y_{ij} = 1 \mid \boldsymbol{\theta}_j)}.
}
\description{
Computes the probability of a correct response for each person--item
combination under the multidimensional 1PL--4PL models. This is the
core IRF (item response function) used throughout the package.
}
\seealso{
\code{\link{fit.MIRT}} for the full model specification.
}
\keyword{internal}
