% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3vcov.R
\name{nobs}
\alias{nobs}
\alias{nobs.MIRT}
\alias{nobs.MGPCM}
\alias{nobs.MGGUM}
\alias{nobs.FCMIRT}
\alias{nobs.FCDCM}
\alias{nobs.FCGDINA}
\alias{nobs.FCGGUM}
\alias{nobs.TIRT}
\title{S3 Methods: nobs}
\usage{
\method{nobs}{MIRT}(object, ...)

\method{nobs}{MGPCM}(object, ...)

\method{nobs}{MGGUM}(object, ...)

\method{nobs}{FCMIRT}(object, ...)

\method{nobs}{FCDCM}(object, ...)

\method{nobs}{FCGDINA}(object, ...)

\method{nobs}{FCGGUM}(object, ...)

\method{nobs}{TIRT}(object, ...)
}
\arguments{
\item{object}{A fitted model object.}

\item{...}{Additional arguments (currently ignored).}
}
\value{
An integer: the sample size \eqn{N}.
}
\description{
Extracts the number of observations (persons) from a fitted
\pkg{ForceChoice} model object. Returns the number of rows in the
response matrix used for model fitting.
}
\section{Methods (by class)}{
\itemize{
\item \code{nobs(MIRT)}: Number of observations (persons) for \code{MIRT} objects.

\item \code{nobs(MGPCM)}: Number of observations for \code{MGPCM} objects.

\item \code{nobs(MGGUM)}: Number of observations for \code{MGGUM} objects.

\item \code{nobs(FCMIRT)}: Number of observations for \code{FCMIRT} objects.

\item \code{nobs(FCDCM)}: Number of observations for \code{FCDCM} objects.

\item \code{nobs(FCGDINA)}: Number of observations for \code{FCGDINA} objects.

\item \code{nobs(FCGGUM)}: Number of observations for \code{FCGGUM} objects.

\item \code{nobs(TIRT)}: Number of observations for \code{TIRT} objects.

}}
