% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{normalize_fc_type}
\alias{normalize_fc_type}
\title{Normalise and recycle forced-choice type vector}
\usage{
normalize_fc_type(fc.type, N.block, default = "RANK")
}
\arguments{
\item{fc.type}{Character vector (or \code{NULL}) describing the FC format of
each block.  Recycled to length \code{N.block} and uppercased.}

\item{N.block}{Number of blocks.}

\item{default}{Default FC type used when \code{fc.type} is \code{NULL}.}
}
\value{
A character vector of length \code{N.block}.
}
\description{
Normalise and recycle forced-choice type vector
}
\keyword{internal}
