% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools-mcmc.R
\name{restore_mcmc_env}
\alias{restore_mcmc_env}
\title{Restore MCMC threading environment to its previous state}
\usage{
restore_mcmc_env(old)
}
\arguments{
\item{old}{A list returned by \code{setup_mcmc_env()}.}
}
\value{
Invisibly \code{NULL}.
}
\description{
Restore MCMC threading environment to its previous state
}
\keyword{internal}
