% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{tirt_par_free_mask}
\alias{tirt_par_free_mask}
\title{Free-parameter mask for TIRT models}
\usage{
tirt_par_free_mask(Q.matrix, lambda.free, psi.free)
}
\arguments{
\item{Q.matrix}{I x D Q-matrix.}

\item{lambda.free}{Logical vector of length I.}

\item{psi.free}{Logical vector of length I.}
}
\value{
An I x (D+1) logical matrix.
}
\description{
Free-parameter mask for TIRT models
}
\keyword{internal}
